fix: Initialize the bus only if required
Currently the bus is initialized when the file is imported
and as the bus requires the window to be present it fails for
e.g. testing without a DOM environment.

To fix this the bus is now initialized on the first usage
and while the window is not present only a stub is used.

Signed-off-by: Ferdinand Thiessen <rpm@fthiessen.de>