chore: update node version to include LTS and active
As per https://github.com/nextcloud/standards/issues/5 apps should
suppot Node 22 now (since October).
But as a library we also need to support 1 year older apps (for the 1yr
maintenance e.g. groupware), so we need to support Node 22 and 20.
Also as we do not have any node dependencies but this is a browser
library we already can support the upcoming LTS 24 (current node).

Signed-off-by: Ferdinand Thiessen <opensource@fthiessen.de>